+++ lib/rubocop/rspec/language/node_pattern.rb
@@ -16,7 +16,7 @@ module RuboCop
         def_node_matcher :spec_group?, spec_groups.block_pattern
 
         def_node_matcher :example_group_with_body?, <<-PATTERN
-          (block #{ExampleGroups::ALL.send_pattern} args [!nil?])
+          (block #{ExampleGroups::ALL.send_pattern} args !nil?)
         PATTERN
 
         def_node_matcher :example?, Examples::ALL.block_pattern